refactor: separate relationship attribute renderers
