+++ packages/renderers/lib/src/renderers/request_item_renderer/read_attribute_request_item_renderer.dart
@@ -2,7 +2,7 @@ import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../request_renderer.dart';
-import 'utils/query_renderer.dart';
+import 'widgets/query_renderer.dart';
 
 class ReadAttributeRequestItemRenderer extends StatelessWidget {
   final LocalRequestDVO request;
