+++ packages/renderers/lib/src/renderers/request_item_renderer/request_item_renderer.dart
@@ -13,7 +13,6 @@ class RequestItemRenderer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    print(item.runtimeType);
     if (item.type.startsWith('Decidable')) return const Text('Decidable');
 
     return switch (item) {
