+++ packages/renderers/lib/src/renderers/request_item_renderer/widgets/identity_attribute_renderer.dart
@@ -1,0 +1,24 @@
+import 'package:enmeshed_types/enmeshed_types.dart';
+import 'package:flutter/material.dart';
+
+class DraftIdentityAttributeRenderer extends StatelessWidget {
+  final DraftIdentityAttributeDVO attribute;
+
+  const DraftIdentityAttributeRenderer({super.key, required this.attribute});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
+
+class IdentityAttributeRenderer extends StatelessWidget {
+  final IdentityAttribute attribute;
+
+  const IdentityAttributeRenderer({super.key, required this.attribute});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
